scientific-skills/multi-panel-figure-assembler/requirements.txt
Pillow>=9.0.0
numpy>=1.21.0
